+++ saleor/graphql/core/mutations.py
@@ -156,16 +156,12 @@ class BaseMutation(graphene.Mutation):
             return None
 
         try:
-            if only_type is not None:
-                _type, pk = from_global_id_or_error(node_id, only_type, field=field)
-            else:
-                # FIXME: warn when supplied only_type is None?
-                only_type, pk = from_global_id_or_error(node_id)
+            object_type, pk = from_global_id_or_error(node_id, only_type, field=field)
 
-            if isinstance(only_type, str):
-                only_type = info.schema.get_type(only_type).graphene_type
+            if isinstance(object_type, str):
+                object_type = info.schema.get_type(object_type).graphene_type
 
-            node = cls.get_node_by_pk(info, graphene_type=only_type, pk=pk, qs=qs)
+            node = cls.get_node_by_pk(info, graphene_type=object_type, pk=pk, qs=qs)
         except (AssertionError, GraphQLError) as e:
             raise ValidationError(
                 {field: ValidationError(str(e), code="graphql_error")}